refactor(features): migrate CleanFeatureExtractor to main extractor module
- Moved the `CleanFeatureExtractor` and `ExtractedFeatures` classes from `extractor_clean.py` to `extractor.py`, consolidating feature extraction logic under a single module.
- Removed the now-obsolete `extractor_clean.py` file, streamlining the codebase and adhering to Clean Architecture principles.
- Updated the constructor to require dependencies explicitly, enhancing clarity and maintainability.
- This refactoring improves the organization of feature extraction components and sets the stage for future enhancements.
